Add pagination for the bulk search results
The bulk search can return hundreds of results and currently the UI
shows them all in a long table.

We want to make this a better UX by introducing pagination.

In order to make this new UX work, we need to store the API response in
a table as the API itself isn't paginated.

To avoid making the same API request over and over again to paginate, we
store the response and set an expiration value. This is effectively a
simple caching system.