Ensure to extend the expiry of a bulk search response
We want the bulk search response to live for as long as someone needs to
access it.

The approach is based on a typical session window where 30 minutes of
inactivity will cause the response to expire.

There is a bug in the implementation at the moment where response
always expires 30 minutes after it was originally created rather than 30
minutes from when it was last accessed.