Add tabbed pagination
The GOVUK pagination component doesn't support being rendered in a tab.

I decided to add a new component that wraps the current GOVUK component
but overrides the link rendering to include a fragment.

This allows the links to use the current tab as part of the URL and
navigate the pagination without affecting the tab value.