Redirect to sign in from the start page
There is a request to make the guidance page the start page.

Navigating to the service should redirect the user to the sign page
without having to press the Start now button again.

I considered updating the guidance page and making the POST request from
there.

This introduces a longer development cycle and makes it slower to make
changes as another team has to be involved.

Instead, I opted to contain the change within this service. There is a
library called repost that can help us here.

It allows us to do a POST redirect by rendering a form with the redirect
URL and auto-submits it.

This has the end result of doing a POST redirect.

Some limitations are that the redirect won't trigger when Javascript is
disabled.